+++ src/main/java/Team_Mute/back_end/domain/reservation_admin/dto/response/ReservationDetailResponseDto.java
@@ -14,6 +14,6 @@ public record ReservationDetailResponseDto(
 	String reservationStatusName,
 	boolean isApprovable,
 	boolean isRejectable,
-	Long statusId
+	Integer statusId
 ) {
 }
