+++ src/main/java/Team_Mute/back_end/domain/reservation_admin/repository/AdminReservationStatusRepository.java
@@ -1,10 +1,11 @@
 package Team_Mute.back_end.domain.reservation_admin.repository;
 
-import Team_Mute.back_end.domain.reservation.entity.ReservationStatus;
-
 import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminReservationStatusRepository extends JpaRepository<ReservationStatus, Long> {
+import Team_Mute.back_end.domain.reservation.entity.ReservationStatus;
+
+public interface AdminReservationStatusRepository extends JpaRepository<ReservationStatus, Integer> {
 	Optional<ReservationStatus> findByReservationStatusName(String reservationStatusName);
 }
